Custom folder structure: verification residence for Madrid
