+++ app/models/verification/residence.rb
@@ -16,8 +16,6 @@ class Verification::Residence
 
   validate :allowed_age
   validate :document_number_uniqueness
-  validate :postal_code_in_madrid
-  validate :residence_in_madrid
 
   def initialize(attrs={})
     self.date_of_birth = parse_date('date_of_birth', attrs)
@@ -43,20 +41,6 @@ class Verification::Residence
 
   def document_number_uniqueness
     errors.add(:document_number, I18n.t('errors.messages.taken')) if User.where(document_number: document_number).any?
-  end
-
-  def postal_code_in_madrid
-    errors.add(:postal_code, I18n.t('verification.residence.new.error_not_allowed_postal_code')) unless valid_postal_code?
-  end
-
-  def residence_in_madrid
-    return if errors.any?
-
-    unless residency_valid?
-      errors.add(:residence_in_madrid, false)
-      store_failed_attempt
-      Lock.increase_tries(user)
-    end
   end
 
   def store_failed_attempt
@@ -95,10 +79,6 @@ class Verification::Residence
 
     def clean_document_number
       self.document_number = self.document_number.gsub(/[^a-z0-9]+/i, "").upcase unless self.document_number.blank?
-    end
-
-    def valid_postal_code?
-      postal_code =~ /^280/
     end
 
 end
