+++ spec/models/residence_spec.rb
@@ -30,29 +30,6 @@ describe Verification::Residence do
       expect(residence.errors[:date_of_birth]).to include("You must be at least 16 years old")
     end
 
-    describe "postal code" do
-      it "should be valid with postal codes starting with 280" do
-        residence.postal_code = "28012"
-        residence.valid?
-        expect(residence.errors[:postal_code].size).to eq(0)
-
-        residence.postal_code = "28023"
-        residence.valid?
-        expect(residence.errors[:postal_code].size).to eq(0)
-      end
-
-      it "should not be valid with postal codes not starting with 280" do
-        residence.postal_code = "12345"
-        residence.valid?
-        expect(residence.errors[:postal_code].size).to eq(1)
-
-        residence.postal_code = "13280"
-        residence.valid?
-        expect(residence.errors[:postal_code].size).to eq(1)
-        expect(residence.errors[:postal_code]).to include("In order to be verified, you must be registered in the municipality of Madrid.")
-      end
-    end
-
     it "should validate uniquness of document_number" do
       user = create(:user)
       residence.user = user
